Move code to make the tests happy
